+++ app/admin/components/cms/cmsDirective.js
@@ -83,7 +83,7 @@
             return {
                 restrict: 'E',
                 replace: true,
-                templateUrl: 'admin/components/cms.html',
+                templateUrl: 'admin/components/cms/cms.html',
                 scope: {},
                 bindToController: {
                     //admin: '='